Big update
Now there is a function that plots model prediction. It takes 9 random images and shows to the model the model predicts which number it is its useful because we can see if the model is actually good. There is also a new confusion metrix plot function. The code is also cleaner.